+++ src/Components/Formularios/LoginForm/LoginForm.jsx
@@ -1,36 +1,35 @@
 import React from "react";
 import { useFormik } from "formik";
 import * as yup from "yup";
-
 import { TextField, Button, Typography, Container } from "@mui/material";
 
-function Form({ formik, validationSchema }) {
+
+const FormContainer = () => {
+  const validationSchema = yup.object({
+    email: yup
+      .string("Ingrese su email")
+      .email("Ingrese un email valido")
+      .required("Email es requerido"),
+    password: yup
+      .string("Ingrese su nombre contraseña")
+      .min(8, "Las contraseña deben tener minimo de 8 caracteres")
+      .required("Contraseña es requerido"),
+  });
+
+  const formik = useFormik({
+    initialValues: {
+      email: "",
+      password: "",
+    },
+    validationSchema: validationSchema,
+    onSubmit: (values) => {
+      console.log(values);
+    },
+  });
+
   return (
     <Container>
       <form onSubmit={formik.handleSubmit}>
-        <TextField
-          fullWidth
-          id="firstName"
-          name="firstName"
-          label="Nombre"
-          value={formik.values.firstName}
-          onChange={formik.handleChange}
-          error={formik.touched.firstName && Boolean(formik.errors.firstName)}
-          helperText={formik.touched.firstName && formik.errors.firstName}
-          style={{ marginBottom: "1em" }}
-        />
-        <TextField
-          fullWidth
-          id="lastName"
-          name="lastName"
-          label="Apellido"
-          value={formik.values.lastName}
-          onChange={formik.handleChange}
-          error={formik.touched.lastName && Boolean(formik.errors.lastName)}
-          helperText={formik.touched.lastName && formik.errors.lastName}
-          style={{ marginBottom: "1em" }}
-        />
-
         <TextField
           fullWidth
           id="email"
@@ -54,13 +53,12 @@ function Form({ formik, validationSchema }) {
           helperText={formik.touched.password && formik.errors.password}
           style={{ marginBottom: "1em" }}
         />
-        {/* Aca deberiamos agregar el componente de Image Upload */}
         <Button color="primary" variant="contained" fullWidth type="submit">
-          Registrarse
+          Ingresa
         </Button>
       </form>
     </Container>
   );
-}
+};
 
-export default Form;
+export default FormContainer;
