+++ src/Pages/Register.jsx
@@ -1,6 +1,8 @@
 import React, { useState } from "react";
 
-import Form from "../Components/Formularios/UserForm/FormContainer";
+import UserForm from "../Components/Formularios/UserForm/UserForm";
+import LoginForm from "../Components/Formularios/LoginForm/LoginForm";
+
 import { Container, Box, Typography } from "@mui/material";
 import { Link } from "react-router-dom";
 
@@ -12,11 +14,37 @@ function Register() {
   const [isRegister, SetIsRegister] = useState(true);
 
   return (
-    <TemplateFormPage
-      Form={Form}
-      ImageBG={loginWallet}
-      isRegister={isRegister}
-    />
+    <Box
+      style={{
+        display: "grid",
+        gridTemplateColumns: "1fr 1fr",
+        placeContent: "center",
+      }}
+    >
+      <Box
+        style={{
+          backgroundImage: `url(${loginWallet})`,
+          height: "100vh",
+          backgroundSize: "cover",
+        }}
+      ></Box>
+      <Box
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          justifyContent: "space-evenly",
+          alignItems: "center",
+        }}
+      >
+        <Typography variant="book1" style={{ marginBottom: "1em" }}>
+          Ingrese tus datos para registrarte como usuario
+        </Typography>
+        <UserForm />
+        <Typography variant="book1">
+          ¿Ya tienes cuenta? <Link to="/login">Login</Link>
+        </Typography>
+      </Box>
+    </Box>
   );
 }
